+++ core/settings.py
@@ -40,8 +40,8 @@ DATABASES = {
     'default': {
         'ENGINE': 'django.db.backends.postgresql',
         'NAME': os.environ.get('POSTGRES_DB', 'guitars'),
-        'USER': os.environ.get('POSTGRES_USER', 'postgres'),
-        'PASSWORD': os.environ.get('POSTGRES_PASSWORD', 'postgres'),
+        'USER': os.environ.get('POSTGRES_USER', 'guitars'),
+        'PASSWORD': os.environ.get('POSTGRES_PASSWORD', 'guitars'),
         'HOST': os.environ.get('POSTGRES_HOST', 'localhost'),
         'PORT': os.environ.get('POSTGRES_PORT', '4455'),
     }